[scheduler] add deterministic component scheduler
This scheduler aims at providing deterministic results, especially in
the presence of recursion cycles. This is based on the locking mechanism
found in RestartScheduler, except that we schedule strongly connected
components instead of lone procedures. Since a cycle cannot span
multiple SCC and each schedule SCC is analyzed on one job, this
alleviates most of the non-determinism caused by the analysis of
recursion cycles. It also contains a mechanism for handling dynamically
discovered recursion cycles (which tend to happen with pulse) in order
to resolve them deterministically.